Refact: Вынес логику карточки в hook и utils
